move WbwTest from Home
